Move delegation E2E tests into focused suite
